+++ tests/test_order.py
@@ -8,40 +8,9 @@ currentdir = os.path.dirname(os.path.abspath(
 parentdir = os.path.dirname(currentdir)
 sys.path.insert(0, parentdir)
 
-from app.app import create_app
-from app.v1.models.userModel import User
-from app.v1.models.orderModel import Order
-user = User()
-order = Order()
+from tests.test_base import TestSetup
 
-class TestOrder(unittest.TestCase):
-
-    """ This class represents the order test case"""
-
-    def setUp(self):
-        self.app = create_app("testing")
-        self.client = self.app.test_client()
-
-        self.add_order = {
-            "user": "Galgallo",
-                "cart": {"burger": 2}
-        }
-        self.new_user = {
-            "username": "test2",
-            "email": "test2@example.com",
-            "password": "test2pass"
-        }
-        self.register = self.client.post(
-            '/auth/register',
-            data=json.dumps(self.new_user),
-            content_type='application/json')
-
-        self.login = self.client.post(
-            '/auth/login',
-            data=json.dumps(self.new_user),
-            content_type='application/json')
-        self.data = json.loads(self.login.data.decode("UTF-8"))
-        self.token = self.data['token']
+class TestOrder(TestSetup):
 
     def test_order_placement(self):
         """ Test new order can be added """
@@ -118,13 +87,8 @@ class TestOrder(unittest.TestCase):
             {'x-access-token': self.token})
         self.assertEqual(res.status_code, 200)
         msg = json.loads(res.data.decode("UTF-8"))
-        self.assertIn("Order updated successfully", msg['message'])
+        self.assertIn("Order status updated successfully", msg['message'])
         
-    def tearDown(self):
-        self.users = user.users
-        self.orders = order.orders
-        self.users.clear()
-        self.orders.clear()
 
 if __name__ == '__main__':
     unittest.main()
